Integrate legacy crypto libraries
Added dependency to the library `ConcordiumWalletCrypto` and added tests for invoking the `parameter_to_json` function. The library is currently used in the iOS reference wallet.

This is only meant as a POC for integrating a Rust library that's already known to work. The dependency will be replaced with the new Rust library that's currently being built.