Fix access modifiers (#7)
Add access modifier public to the appropriate types etc. to ensure that the symbols in the library that should be available from applications using the library are in fact so. The modifiers were added incrementally while using the library in an external tool to ensure that they were in fact all necessary.