Replace example tool and gRPC client "tests" with CLI
The CLI is able to invoke all currently defined methods of `ConcordiumNodeClient`.

The unit tests (which didn't actually perform any assertions) have been replaced with a script `test.sh` inside the CLI folder which executes all the same calls as these tests.

This removes the need for an accessible node when running unit tests, meaning that we can start running `swift test` as part of the CI workflow.