Integrate Rust bindings as library (#11)
Add Rust binding library `ConcordiumWalletCrypto` which implements the functions currently exposed in the `ConcordiumHdWallet` in the JavaScript SDK. The functions are made available from the SDK via `ConcordiumHdWallet` as well. The unit tests have been ported from the other SDK also.

The bindings are generated by [UniFFI](https://mozilla.github.io/uniffi-rs/Overview.html) via [`cargo-swift`](https://github.com/antoniusnaumann/cargo-swift/). That tool was mainly helpful for setting up the structure of the project, it probably isn't really required for actually building the lib. As it's ability to compile seems a little flaky and because we might want to build fewer variants for the purpose of CI, it seems reasonable that we'd replace this dependency with a simple script that does the same thing.

Resolves https://concordium.atlassian.net/browse/CBW-1552.

References:
- JS SDK: https://github.com/Concordium/concordium-node-sdk-js/blob/main/packages/sdk/src/wasm/HdWallet.ts
- Unit tests ported: https://github.com/Concordium/concordium-node-sdk-js/blob/main/packages/sdk/test/ci/HdWallet.test.ts#L186