Replace cargo-swift with equivalent script (#15)
Building the framework with a shell script will (hopefully) make it clearer what's actually going on.

Going forward it will also give us more control over the exact procedure like for instance whether to build universal binaries or support arm64 only. For example it's only necessary to build a single target on the CI when running unit tests (for now we build for all platforms on the CI).

Last but not least it will allow us to remove a dependency that we don't actually need.