SwiftPM: Add ability to use local crypto library
Added processing of build-time environment variable `CONCORDIUM_WALLET_CRYPTO_PATH` which may be used to select a local checkout of the crypto library (and a local binary framework compiled from it) during development.

The readme was also updated with accurate information on `ConcordiumWalletCrypto` after the changes in #2 (5c06fa29803cdf4c07135a1f514849d5bec19652) where this was missed.