Example CLI: Fix name and cleanup
When working on the next features, we ran into a [known bug in Swift](https://github.com/apple/swift/issues/55127) that prevents the annotation `@main` from being usable from within a file named `main.swift`.

According to a [note in the docs of `ArgumentParser`](https://apple.github.io/swift-argument-parser/documentation/argumentparser/gettingstarted/) (and apparently not documented anywhere else), this shouldn't work at all:

> The Swift compiler uses either the type marked with @main or a main.swift file as the entry point for an executable program. You can use either one, but not both [...]

So it's a bit of a mystery why it works in the current version...

To fix the problem we rename `main.swift` to `CLI.swift`.

Also, the executable is renamed to `concordium-example-client` to ensure that it never is mistaken for a production ready tool.

Finally, some variables were renamed for consistency and the readme got an overhaul.