Example CLI: Downgrade swift-tools-version
The used version (5.9) doesn't work on my older Intel Mac, which was a bit annoying. The only change required for downgrading was re-adding the intermediate directory that was removed in #22 (commit 687e76d674e90a1b1d27c03d38ed0904ebc2343e).

Also, the dependency on the SDK was changed to a simple local dependency on path `../..`. The env var override turned out too annoying in practice, in particular because it causes changes to `Package.resolved` that may easily get committed by accident.