Implement identity issuance and recovery (#23)
The `NodeClient` protocol is extended with methods for querying IPs and ARs and a method for sending a credential deployment. The IP info exposed via gRPC doesn't include the URLs for starting identity issuance and recovery, so Wallet Proxy integration is added for use when those are needed.

The crypto functions for construction identity/credential requests added in https://github.com/Concordium/concordium-wallet-crypto-swift/pull/6 are integrated via `WalletSeed`. The credential deployment serialization functions are integrated in `Transaction.swift`.

URLs to invoke the IPs are conveniently built using helpers in `WalletIdentity.swift`. Due to the fact that IPs respond with identity objects etc. encoded as JSON, the relevant types from `ConcordiumWalletCrypto` are extended to conform with `Decodable`. A convenient helper for building the request payloads is included as well.

The example CLI is extended to support all the new functionality via the commands:

- `wallet identity issue` (example: `concordium-example-client wallet --seed-phrase="..." --identity-provider-id=1 --identity-index=2 identity issue`)
- `wallet identity recover` (example: `concordium-example-client wallet --seed-phrase="..." --identity-provider-id=1 --identity-index=2 identity recover`)
- `wallet identity create-account` (example: `concordium-example-client wallet --seed-phrase="..." --identity-provider-id=1 --identity-index=2 identity create-account --credential-counter=0`)

The readme includes all relevant details.