CI: Run tests on all supported macOS runners (#26)
The lint/test job now runs on all [available](https://github.com/actions/runner-images/?tab=readme-ov-file#available-images) ("normal"-sized) macOS runners of OS versions that the SDK builds on.

In this regard, we now explicitly support building on macOS 12+, so we don't use the `macos-11` runner. This runner only supports swift-tools up to version 5.5 and we use 5.6 (downgrading would take away ability to depend on SwiftFormat by exact version).

The formatting check now only runs in a single job. This was initially done to make macOS 11 work because its version of SwiftFormat didn't support all the options in the used command, but it made sense to keep in anyway.

Finally, a check that the example CLI builds successfully was added to the job.